Function manycast::cli::create_schedule_task

fn create_schedule_task(
    source_address: IP,
    destination_addresses: Vec<Address>,
    task_type: u32,
    rate: u32,
    client_ids: Vec<u32>,
    unicast: bool,
    ipv6: bool,
    traceroute: bool
) -> ScheduleTask

Create a Verfploeter ScheduleTask message that can be sent to the server.

Arguments

‘source_address’ - the source address to be used for this task (will be overwritten by the clients if they have a source address specified locally)

‘destination_addresses’ - a vector of destination addresses that will be probed in this task (e.g., the hitlist)

‘task_type’ - the type of task, can be 1: ICMP/ping, 2: UDP/A, 3: TCP, 4: UDP/CHAOS

‘rate’ - the rate (packets / second) at which clients will send out probes (default: 1000)

‘client_ids’ - the list of clients (by IDs) who participate in the measurement (empty list means all clients participate)

‘unicast’ - a boolean that determines whether the clients must use their unicast source address

‘ipv6’ - a boolean that determines whether the addresses are IPv6 or not // TODO useless ?

‘traceroute’ - a boolean that determines whether the clients must perform traceroute measurements

Examples

let task = create_schedule_task(124.0.0.0, vec![1.1.1.1, 8.8.8.8], 1, 1400, vec![]);