Struct manycast::cli::CliClient

pub struct CliClient {
    grpc_client: ControllerClient<Channel>,
}

A CLI client that creates a connection with the ‘server’ and sends the desired commands based on the command-line input.

Fields
grpc_client: ControllerClient<Channel>
Implementations

impl CliClient

async fn do_task_to_server(
&mut self,
task: ScheduleTask,
task_type: u32,
cli: bool
) -> Result<(), Box<dyn Error>>

Send the ‘do_task’ command to the server, await the task results, and print them out to command-line & file as CSV.

Arguments

‘task’ - the task that is being sent to the server

‘task_type’ - the type of task that is being sent, and the type of the task results we will receive

‘cli’ - a boolean that determines whether the results should be printed to the command-line (will be true if –stream was added to the start command)

async fn list_clients_to_server(&mut self) -> Result<(), Box<dyn Error>>

Sends a list clients command to the server, awaits the result, and prints it to command-line.