Module manycast::client::inbound

Functions

get_pcap ğŸ”’

Create a pcap capture object with the given filter.

handle_results ğŸ”’

Thread for handling the received replies, wrapping them in a TaskResult, and streaming them back to the main client class.

listen_ping

Listen for incoming ping/ICMP packets, these packets must have our payload to be considered valid replies.

listen_tcp

Listen for incoming TCP/RST packets, these packets must have the correct destination port, have the right flags set (RST), and have ACK == 0 for it to be considered a reply.

listen_udp

Listen for incoming UDP DNS packets, these packets must have a DNS A record reply and use the correct port numbers to be considered a reply.

parse_chaos ğŸ”’

Attempts to parse the DNS Chaos record from a UDP payload body.

parse_dns_a_record ğŸ”’

Attempts to parse the DNS A record from a UDP payload body.

parse_icmp_dest_unreachable ğŸ”’

ICMP Destination unreachable parser, used for DNS A record probing

parse_icmp_time_exceeded ğŸ”’

ICMP Time exceeded parser, used for traceroute

parse_icmpv4 ğŸ”’

Parse ICMPv4 packets (including v4 headers) into a VerfploeterResult.

parse_icmpv6 ğŸ”’

Parse ICMPv6 packets (including v6 headers) into a VerfploeterResult.

parse_ipv4 ğŸ”’

Parse packet bytes into an IPv4 header, returns the IP result for this header and the payload.

parse_ipv6 ğŸ”’

Parse packet bytes into an IPv6 header, returns the IP result for this header and the payload.

parse_tcp ğŸ”’

Parse TCP body into a VerfploeterResult.

parse_tcpv4 ğŸ”’

Parse TCPv4 packets (including v4 headers) into a VerfploeterResult.

parse_tcpv6 ğŸ”’

Parse TCPv6 packets (including v6 headers) into a VerfploeterResult.

parse_udpv4 ğŸ”’

Parse UDPv4 packets (including v4 headers) into a VerfploeterResult.

parse_udpv6 ğŸ”’

Parse UDPv6 packets (including v6 headers) into a VerfploeterResult.