Function manycast::client::inbound::handle_results

fn handle_results(
    metadata: Metadata,
    tx: &UnboundedSender<TaskResult>,
    tx_f: Sender<()>,
    task_id: u32,
    client_id: u8,
    result_queue_sender: Arc<Mutex<Option<Vec<VerfploeterResult>>>>
)

Thread for handling the received replies, wrapping them in a TaskResult, and streaming them back to the main client class.

Arguments

‘metadata’ - contains the metadata of this listening client (the hostname)

‘tx’ - sender to put task results in

‘tx_f’ - channel that gets closed when the outbound prober is finished for the current measurement

‘task_id’ - the task_id of the current measurement

‘client_id’ - the unique client ID of this client

‘result_queue_sender’ - contains a vector of all received replies as VerfploeterResult