Function manycast::client::inbound::listen_ping

pub fn listen_ping(
    metadata: Metadata,
    socket: Arc<Socket>,
    tx: UnboundedSender<TaskResult>,
    tx_f: Sender<()>,
    task_id: u32,
    client_id: u8
)

Listen for incoming ping/ICMP packets, these packets must have our payload to be considered valid replies.

Creates two threads, one that listens on the socket and another that forwards results to the server and shuts down the receiving socket when appropriate.

For a received packet to be considered a reply, it must be an ICMP packet with the current task ID in the first 4 bytes of the ICMP payload. From these replies it creates task results that are put in a result queue, which get sent to the server.

Arguments

‘metadata’ - contains the metadata of this listening client (the hostname and client ID)

‘socket’ - the socket to listen on

‘tx’ - sender to put task results in

‘tx_f’ - channel that gets closed when the outbound prober is finished for the current measurement

‘task_id’ - the task_id of the current measurement

‘client_id’ - the unique client ID of this client