Function manycast::client::outbound::perform_udp

pub fn perform_udp(
    socket: Arc<Socket>,
    rx_f: Receiver<()>,
    client_id: u8,
    source_address: u32,
    source_port: u16,
    outbound_channel_rx: Receiver<Task>,
    finish_rx: Receiver<()>,
    rate: u32
)

Performs a UDP DNS task by sending out DNS A Record requests with a custom domain name.

This domain name contains the transmission time, the client ID of the prober, the task ID of the current task, and the source and destination address of the probe.

We notify the receiver that it is finished 10 seconds after the last packet is sent.

Arguments

‘socket’ - the socket to send the probes from

‘rx_f’ - channel that we close when we are done probing, to notify the inbound listener that we are finished

‘client_id’ - the unique client ID of this client

‘source_addr’ - the source address we use in our probes

‘source_port’ - the source port we use in our probes

‘outbound_channel_rx’ - on this channel we receive future tasks that are part of the current measurement

‘finish_rx’ - used to abort the measurement

‘rate’ - the number of probes to send out each second