Struct manycast::client::Client

pub struct Client {
    grpc_client: ControllerClient<Channel>,
    metadata: Metadata,
    source_address: IP,
    source_port: u16,
    dest_port: u16,
    active: Arc<Mutex<bool>>,
    current_task: Arc<Mutex<u32>>,
    outbound_tx: Option<Sender<Data>>,
    inbound_tx_f: Option<Vec<Sender<()>>>,
    multi_probing: bool,
}

The client that is ran at the anycast sites and performs tasks as instructed by the server to which it is connected to

Fields

‘grpc_client’ - the client connection with the server

‘metadata’ - used to store this client’s hostname and unique client ID

‘source_address’ - contains the source address this client will use for outgoing probes (optional value that can be defined when creating this client)

‘source_port’ - contains the source port this client will use for outgoing probes (optional value that can be defined when creating this client)

‘dest_port’ - contains the destination port (OR ICMP identifier value) this client will use for outgoing probes (optional value that can be defined when creating this client)

‘active’ - boolean value that is set to true when the client is currently doing a measurement

‘current_task’ - contains the task ID of the current measurement

‘outbound_tx’ - contains the sender of a channel to the outbound prober that tasks are send to

‘inbound_tx_f’ - contains the sender of a channel to the inbound listener that is used to signal the end of a measurement

‘multi_probing’ - boolean value that is set to true when the client is configured to send probes from all configured origins (i.e., address/port combinations)

Fields
grpc_client: ControllerClient<Channel>metadata: Metadatasource_address: IPsource_port: u16dest_port: u16active: Arc<Mutex<bool>>current_task: Arc<Mutex<u32>>outbound_tx: Option<Sender<Data>>inbound_tx_f: Option<Vec<Sender<()>>>multi_probing: bool
Implementations

impl Client

pub async fn new(args: &ArgMatches<'_>) -> Result<Client, Box<dyn Error>>

Create a client instance, which includes establishing a connection with the server.

Extracts the parameters of the command-line arguments.

Arguments

‘args’ - contains the parsed command-line arguments

async fn connect( address: &str ) -> Result<ControllerClient<Channel>, Box<dyn Error>>

Connect to the server.

Arguments

‘address’ - the address of the server in string format, containing both the IPv4 address and port number

Example

let grpc_client = connect("127.0.0.0:50001");

fn init( &mut self, task: Task, client_id: u8, outbound_f: Option<Receiver<()>>, probing: bool, igreedy: bool )

Initialize a new measurement by creating outbound and inbound threads, and ensures tasks are sent back to the server.

Extracts the protocol type from the task, and determines which source address to use. Creates a socket to send out probes and receive replies with, calls the appropriate inbound & outbound functions. Creates an additional thread that forwards task results to the server.

Arguments

‘task’ - the first ‘Task’ message sent by the server for the new measurement

‘client_id’ - the unique ID of this client

‘outbound_f’ - a channel used to send the finish signal to the outbound prober

‘probing’ - a boolean that indicates whether this client has to send out probes

‘igreedy’ - a boolean that indicates whether this client has to use the local unicast address to perform latency measurements

async fn connect_to_server(&mut self) -> Result<(), Box<dyn Error>>

Establish a formal connection with the server.

Obtains a unique client ID from the server, establishes a stream for receiving tasks, and handles tasks as they come in.

async fn get_client_id_to_server(&mut self) -> Result<ClientId, Box<dyn Error>>

Send the get_client_id command to the server to obtain a unique client ID

async fn send_result_to_server( &mut self, task_result: TaskResult ) -> Result<(), Box<dyn Error>>

Send a TaskResult to the server

async fn task_finished_to_server( &mut self, finished: Finished ) -> Result<(), Box<dyn Error>>

Let the server know the current measurement is finished.

When a measurement is finished the server knows not to expect any more results from this client.

Arguments

‘finished’ - the ‘Finished’ message to send to the server