Module manycast::custom_module::verfploeter

Modules

address

Nested message and enum types in Address.

controller_client

Generated client implementations.

controller_server

Generated server implementations.

ip_result

Nested message and enum types in IPResult.

schedule_task

Nested message and enum types in ScheduleTask.

task

Nested message and enum types in Task.

trace_result

Nested message and enum types in TraceResult.

udp_payload

Nested message and enum types in UDPPayload.

verfploeter_result

Nested message and enum types in VerfploeterResult.

Structs

Ack

Acknowledgement of a task, with a boolean whether it was successful or not, and a possible error message.

Address

Address message, can be either an IPv4 or IPv6 address

Client

A client, which contains a client ID, and metadata

ClientId

A client ID

ClientList

A list of Client (an empty list is used when all clients have to send out probes)

DnsARecord

UDP payload, which is extracted from the domain of the DNS A Record reply, containing information about the probe

DnsChaos

CHAOS payload, which is extracted from the domain of the DNS CHAOS reply

Empty

Empty message

End

An end message, indicates the client should stop sending out probes

Finished

TaskFinished message, sent by a client to the server when a task is finished

IPv4Result

IPv4Result, contains all information extracted from the IPv4 header: source, destination addresses and TTL

IPv6

IPv6 address message (since protobuf does not support 128 bit values)

IPv6Result

IPv6Result, contains all information extracted from the IPv6 header: source, destination addresses

IpResult

IPResult, contains either an IPv4Result or an IPv6Result, and the TTL

Metadata

Metadata (as used in Client), contains a hostname, and the origin the client uses

Origin

Origin: the source address and port combinations used by clients in a task

Ping

A Ping, which has a list of destination addresses

PingPayload

PingPayload. The payload of a ping which contains the transmission time, source/dest address, and the sender client ID

PingResult

Result of a ping which has a receive time, an ipv4 result, and the ping payload information

ScheduleTask

Schedule a task with the Server can be either a Ping, UDP, or TCP task, with a probing rate, source address, and a list of clients that have to send out probes

Start

A start message, indicates the probing rate (packets per second), whether the client should send out probes, the type of task, and the source address to be used

Task

A task for a Client, can be either a Start, Ping, UDP, TCP, or an Empty task

TaskResult

A TaskResult, contains the Client that executed the task, and a list of results

Tcp

TCP, which has a list of destination addresses

TcpResult

TCP result which contains a receive time, an ipv4 result, source & destination ports, and the ack & seq numbers

Trace

A traceroute task, which has a destination address, a max_ttl (expected number of hops to traceroute), and a list of flows (a unique flow for each receiving client)

TraceResult

TraceResult, contains the IP result, TTL, receive time, transmit time, and the sender client ID

Udp

UDP, which has a list of destination addresses

UdpPayload

UDP payload, which contains the transmission time, source/dest address, and the sender client ID

UdpResult

UDP result contains the receive time, source and destination port, an ipv4 result, the UDP payload, and a code (0-15 for ICMP codes, 16 => UDP response)

VerfploeterResult

A result, either a PingResult, UDPResult, or TCPResult