Trait manycast::custom_module::verfploeter::controller_server::Controller

pub trait Controller: Send + Sync + 'static {
    type ClientConnectStream: Stream<Item = Result<Task, Status>> + Send + 'static;
    type DoTaskStream: Stream<Item = Result<TaskResult, Status>> + Send + 'static;

    // Required methods
    fn task_finished<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Finished>
    ) -> Pin<Box<dyn Future<Output = Result<Response<Ack>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn client_connect<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Metadata>
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::ClientConnectStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_task<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ScheduleTask>
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::DoTaskStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn list_clients<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Empty>
    ) -> Pin<Box<dyn Future<Output = Result<Response<ClientList>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn send_result<'life0, 'async_trait>(
        &'life0 self,
        request: Request<TaskResult>
    ) -> Pin<Box<dyn Future<Output = Result<Response<Ack>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_client_id<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Metadata>
    ) -> Pin<Box<dyn Future<Output = Result<Response<ClientId>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with ControllerServer.

Required Associated Types

type ClientConnectStream: Stream<Item = Result<Task, Status>> + Send + 'static

Server streaming response type for the ClientConnect method.

type DoTaskStream: Stream<Item = Result<TaskResult, Status>> + Send + 'static

Server streaming response type for the DoTask method.

Required Methods

fn task_finished<'life0, 'async_trait>( &'life0 self, request: Request<Finished> ) -> Pin<Box<dyn Future<Output = Result<Response<Ack>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Client lets the server know the task is finished

fn client_connect<'life0, 'async_trait>( &'life0 self, request: Request<Metadata> ) -> Pin<Box<dyn Future<Output = Result<Response<Self::ClientConnectStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Client announces itself, and starts its connection with the Server. The server sends a stream of Task

fn do_task<'life0, 'async_trait>( &'life0 self, request: Request<ScheduleTask> ) -> Pin<Box<dyn Future<Output = Result<Response<Self::DoTaskStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

CLI instructs Server to perform a measurement

fn list_clients<'life0, 'async_trait>( &'life0 self, request: Request<Empty> ) -> Pin<Box<dyn Future<Output = Result<Response<ClientList>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

CLI asks Server to list clients, Server returns ClientList

fn send_result<'life0, 'async_trait>( &'life0 self, request: Request<TaskResult> ) -> Pin<Box<dyn Future<Output = Result<Response<Ack>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Client sends TaskResult to Server

fn get_client_id<'life0, 'async_trait>( &'life0 self, request: Request<Metadata> ) -> Pin<Box<dyn Future<Output = Result<Response<ClientId>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Client obtains a unique client ID from the server