Module manycast::net

Structs

DNSARecord

A DNS A record

ICMP4Packet

An ICMP4Packet (ping packet) https://en.wikipedia.org/wiki/Internet_Control_Message_Protocol#header_rest

IPv4Packet

A struct detailing an IPv4Packet https://en.wikipedia.org/wiki/Internet_Protocol_version_4

PseudoHeader

Struct defining a pseudo header that is used by both TCP and UDP to calculate their checksum

TCPPacket

A TCPPacket https://en.wikipedia.org/wiki/Transmission_Control_Protocol

UDPPacket

An UDPPacket (UDP packet) https://en.wikipedia.org/wiki/User_Datagram_Protocol

Enums

PacketPayload

Definition of the IPV4Packet payload (either ICMPv4, UDP, TCP, or unimplemented)

Constants

INFO_URL 🔒

Functions

calculate_checksum

Calculate the checksum for a UDP/TCP packet.

read_dns_name 🔒

Read a DNS name that is contained in a DNS response. Returns the domain name string of the A record reply.