Struct manycast::server::ControllerService

pub struct ControllerService {
    clients: Arc<Mutex<ClientList>>,
    senders: Arc<Mutex<Vec<Sender<Result<Task, Status>>>>>,
    cli_sender: Arc<Mutex<Option<Sender<Result<TaskResult, Status>>>>>,
    open_tasks: Arc<Mutex<HashMap<u32, u32>>>,
    current_task_id: Arc<Mutex<u32>>,
    current_client_id: Arc<Mutex<u32>>,
    active: Arc<Mutex<bool>>,
}

Struct for the Server service

Fields

‘clients’ - a ClientList that contains all connected clients (hostname and client ID)

‘senders’ - a list of senders that connect to the clients, these senders are used to stream Tasks

‘cli_sender’ - the sender that connects to the CLI, to stream TaskResults

‘open_tasks’ - a list of the current open tasks, and the number of clients that are currently working on it

‘current_task_id’ - keeps track of the last used task ID and is used to assign a unique task ID to a new measurement

‘current-client_id’ - keeps track of the last used client ID and is used to assign a unique client ID to a new connecting client

‘active’ - a boolean value that is set to true when there is an active measurement

Fields
clients: Arc<Mutex<ClientList>>senders: Arc<Mutex<Vec<Sender<Result<Task, Status>>>>>cli_sender: Arc<Mutex<Option<Sender<Result<TaskResult, Status>>>>>open_tasks: Arc<Mutex<HashMap<u32, u32>>>current_task_id: Arc<Mutex<u32>>current_client_id: Arc<Mutex<u32>>active: Arc<Mutex<bool>>