Function manycast::client::inbound::listen_ping

pub fn listen_ping(
    tx: UnboundedSender<TaskResult>,
    rx_f: Receiver<()>,
    task_id: u32,
    client_id: u8,
    v6: bool,
    filter: String,
    traceroute: bool
)

Listen for incoming ping/ICMP packets, these packets must have our payload to be considered valid replies.

Creates two threads, one that listens on the socket and another that forwards results to the server and shuts down the receiving socket when appropriate.

For a received packet to be considered a reply, it must be an ICMP packet with the current task ID in the first 4 bytes of the ICMP payload. From these replies it creates task results that are put in a result queue, which get sent to the server.

Arguments

‘tx’ - sender to put task results in

‘rx_f’ - channel that is used to signal the end of the measurement

‘task_id’ - the task_id of the current measurement

‘client_id’ - the unique client ID of this client

‘v6’ - whether to parse the packets as IPv6 or IPv4

‘filter’ - the pcap filter to use

‘traceroute’ - whether to handle additional traceroute packets (ICMP time exceeded)