Struct manycast::server::ControllerService

pub struct ControllerService {
    clients: Arc<Mutex<ClientList>>,
    senders: Arc<Mutex<Vec<Sender<Result<Task, Status>>>>>,
    cli_sender: Arc<Mutex<Option<Sender<Result<TaskResult, Status>>>>>,
    open_tasks: Arc<Mutex<HashMap<u32, u32>>>,
    current_task_id: Arc<Mutex<u32>>,
    current_client_id: Arc<Mutex<u32>>,
    active: Arc<Mutex<bool>>,
    traceroute_targets: Arc<Mutex<HashMap<IP, (Vec<u8>, Instant, u8, Vec<Origin>)>>>,
    traceroute: Arc<Mutex<bool>>,
    interval: u64,
}

Struct for the Server service

Fields

‘clients’ - a ClientList that contains all connected clients (hostname and client ID)

‘senders’ - a list of senders that connect to the clients, these senders are used to stream Tasks

‘cli_sender’ - the sender that connects to the CLI, to stream TaskResults

‘open_tasks’ - a list of the current open tasks, and the number of clients that are currently working on it

‘current_task_id’ - keeps track of the last used task ID and is used to assign a unique task ID to a new measurement

‘current_client_id’ - keeps track of the last used client ID and is used to assign a unique client ID to a new connecting client

‘active’ - a boolean value that is set to true when there is an active measurement

‘traceroute_targets’ - a map that keeps track of the clients that have received probe replies for a specific target, and the ‘flows’ that reach each client

‘traceroute’ - a boolean value that is set to true when traceroute is enabled

‘interval’ - the interval between clients sending out probes to the same target

Fields
clients: Arc<Mutex<ClientList>>senders: Arc<Mutex<Vec<Sender<Result<Task, Status>>>>>cli_sender: Arc<Mutex<Option<Sender<Result<TaskResult, Status>>>>>open_tasks: Arc<Mutex<HashMap<u32, u32>>>current_task_id: Arc<Mutex<u32>>current_client_id: Arc<Mutex<u32>>active: Arc<Mutex<bool>>traceroute_targets: Arc<Mutex<HashMap<IP, (Vec<u8>, Instant, u8, Vec<Origin>)>>>traceroute: Arc<Mutex<bool>>interval: u64